split StudySummary from PrivacyPolicy
We had combined the study summary and privacy policy into one component. We're splitting them back up.
They will both be located in js/onboarding.
